+++ records_mover/db/bigquery/loader.py
@@ -2,6 +2,7 @@ from typing import Union, List, IO, Tuple, Optional
 from ...records.hints import complain_on_unhandled_hints
 import pprint
 import sqlalchemy
+from ...records.types import RecordsFormatType
 from ...records.load_plan import RecordsLoadPlan
 from ...records.records_format import (
     BaseRecordsFormat, DelimitedRecordsFormat, ParquetRecordsFormat
@@ -117,3 +118,10 @@ class BigQueryLoader(LoaderFromFileobj,
 
     def known_supported_records_formats_for_load(self) -> List[BaseRecordsFormat]:
         return [DelimitedRecordsFormat(variant='bigquery'), ParquetRecordsFormat()]
+
+    def best_records_format_variant(self, records_format_type: RecordsFormatType) ->\
+            Optional[str]:
+        if records_format_type == 'delimited':
+            return 'bigquery'
+        else:
+            return None
